+++ src/Pim/Bundle/VersioningBundle/PimVersioningBundle.php
@@ -5,8 +5,9 @@ namespace Pim\Bundle\VersioningBundle;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Oro\Bundle\EntityBundle\DependencyInjection\Compiler\DoctrineOrmMappingsPass;
-use Pim\Bundle\VersioningBundle\DependencyInjection\Compiler;
 use Pim\Bundle\CatalogBundle\PimCatalogBundle;
+use Pim\Bundle\TransformBundle\DependencyInjection\Compiler\SerializerPass;
+use Pim\Bundle\VersioningBundle\DependencyInjection\Compiler;
 
 /**
  * Pim Versioning Bundle
@@ -23,7 +24,8 @@ class PimVersioningBundle extends Bundle
     public function build(ContainerBuilder $container)
     {
         $container
-            ->addCompilerPass(new Compiler\RegisterUpdateGuessersPass());
+            ->addCompilerPass(new Compiler\RegisterUpdateGuessersPass())
+            ->addCompilerPass(new SerializerPass('pim_versioning.serializer'));
 
         $versionMappings = [
             realpath(__DIR__ . '/Resources/config/model/doctrine') => 'Pim\Bundle\VersioningBundle\Model'